Fix openstack subordinates using 'cs:'. (#52)
* Fix subordinates using 'cs:'.
* Fix _default_used could be unset.

If subordinate charm does not have channels this means it is either before focal, or using cs:

In any case assuming it ussuri/stable will not break the code because and openstack

if the principal is not (ussuri/victoria/..)/stable we will not try to upgrade it
principals thus the subordinate binaries already upgraded to whatever version we are trying to upgrade
basically if subordinate is cs:charm stable principal is ussuri/stable we will first upgrade principal to victoria than upgrade subordinate to victoria. We will not refresh charm it its origin is unidentified.