Remove OpenStack bundle and make the backup test in the smoke bundle
Currently, COU is using a heavy OpenStack bundle to just test the
backup functionality. It's possible to mock the part that checks
if there is a relation between keystone and mysql and use a lighter
bundle containing just 3 units of mysql-innodb-cluster and 1 unit
of designate-bind.

The upgrade of more complex bundles probably will happen on SolQA
environment (TBD).

Pros:
- faster testing
- less time running the CI
- No need to have a dedicated github runner to run the project
- Same functional test coverage with less time and less resources.

Cons:
- Miss the plan generation and run of more complex bundles.

Closes: #189