Update breaking changes from starter-pack and improve workflow
- There are recent breaking changes in canonical/sphinx-docs-starter-pack
that causes our workflow to fail. Updated those changes in COU to align.
- Update the workflow to only trigger documentation checks if a PR
modifies `docs/`. This will prevent un-necessary runs. I'm keeping the
on.push as is because I think it is good to have a way to notify us
about dependency breakage (those issues are not affecting live doc site).