Feature/bseng 1937/upgrade unit by unit (#237)
Add UnitUpgradeStep and is_unit_upgrade_step, also use sub_steps as
property to prevent defining it as not list of BaseStep's. Raising
TypeError in add_step if object is not based on BaseStep.

Use UnitUpgradeStep, we can upgrade packages in individual step for each
unit and
run these steps in parallel or sequentially. Right now, it's run
sequentially, so no change in previous approach.