Add cache property for applications to COUModel (#275)
Like this it will be possible to access all application in model,
without calling get_applications. This will be used in planning stage,
where we are not running async functions and for any app, which require
using other applications. e.g. ceph-osd need to check nova-compute
status.

With this feature, we can easily access all colocated nova-compute and
ceph-mon during pre- and post-upgrade steps in ceph-osd.