add global and func test concurrency (#301)
Right now we are using global concurrency for workflow name and like
this we are not running even lint and unit tests, which are run on
GitHub runner instead of self-hosted one. That's why I do not see reason
to block them to be run.

We also decided to use official GitHub runner for functional tests
instead of self-hosted runner.