Add logic for step to raise HaltUpgradeExecution (#316)
This feature was introduced in [1], however the logic was not handling
in parallel execution.

After more discussion, we agreed on following execution behavior for

# Sequential:
If any step in sequence raise an exception the sequence will be stopped
immediately, however if the exception is `HaltUpgradeExecution` the
sequence will continue with skipping all dependent steps.
For example, if the `Check if unit nova-compute/0 has no VMs running
before upgrading.` in following example raise `HaltUpgradeExecution` all
steps with prefix `├──` will be skipped and the `Enable nova-compute
scheduler from unit: 'nova-compute/0'.`
```bash
Upgrade plan for 'nova-compute' to victoria
        ...
        Upgrade plan for units: nova-compute/0, nova-compute/1, nova-compute/2
                Upgrade plan for unit: nova-compute/0
                        Disable nova-compute scheduler from unit: 'nova-compute/0'.
                        Check if unit nova-compute/0 has no VMs running before upgrading.
                        ├── Pause the unit: 'nova-compute/0'.
                        ├── Upgrade the unit: 'nova-compute/0'.
                        ├── Resume the unit: 'nova-compute/0'.
                        Enable nova-compute scheduler from unit: 'nova-compute/0'.
        ...
```
In other hand, if `├── Pause the unit: 'nova-compute/0'.` raise some
exception, no further steps will be executed in `Upgrade plan for unit:
nova-compute/0` and COU will exit.

# Parallel
In parallel execution, we are waiting for all steps to finished, even if
any raise an exception. After all steps finished, the exception is
raised with all collected exception form steps.
For example, if `Disable nova-compute scheduler from unit:
'nova-compute/0'.` raise an exception in the following example, the
`Upgrade plan for unit: nova-compute/0` will be stopped, but `Upgrade
plan for unit: nova-compute/1` and `Upgrade plan for unit:
nova-compute/2` will continue until they finished. After each step in
those the steps finished, the `RunUpgradeError` is raised and COU exit.
It means that ` Wait 1800s for model test_model to reach the idle
state.` will not be executed.

```bash
Upgrade plan for 'nova-compute' to victoria
        ...
        Upgrade plan for units: nova-compute/0, nova-compute/1, nova-compute/2
                Upgrade plan for unit: nova-compute/0
                        Disable nova-compute scheduler from unit: 'nova-compute/0'.
                        ...
                Upgrade plan for unit: nova-compute/1
                        ...
                Upgrade plan for unit: nova-compute/2
                        ...
        Wait 1800s for model test_model to reach the idle state.
        Check if the workload of 'nova-compute' has been upgraded
```

---
[1]: https://github.com/canonical/charmed-openstack-upgrader/pull/239