current_os_release also consider OpenStack release set in configutation
Apps that are colocated with nova-compute will upgrade packages
during nova-compute upgrade. To ensure to change the 'source' or
'openstack-origin' and run all the upgrade steps necessary, it's
necessary to include the OpenStack release set in the application
configuration.

This patch also fix the post-upgrade generation to ensure that the
require-osd-release option matches the ceph-osd version by passing
control-plane apps where ceph-mon is and not on data-plane apps

Closes: #399, #396 and #380