Update designate-bind upper versions
Pinning to an upper version at patch level caused cou to start failing
to verify the upgrade, since the workload version of
designate-bind has new patch releases since it was pinned
and now falls outside the original version range.

It should be fine to pin to < 9.19,
since any patch release of 9.18 should be fine and expected.

Fixes: #512