Improve checking for mismatched unit openstack versions
If the units of an application are running mismatched openstack versions,
this is normally ok
(a user may have manually ran an apt upgrade for example).
But if the versions are so mismatched
that they are detected as being for mismatched openstack releases,
then that's a problem,
potentially indicating a failed upgrade on a unit for example.
Eg. if one unit is running a workload version only for victoria,
while the other two units are running a version for wallaby,
this is a consistency problem,
and COU should block an automated upgrade.

COU was previously checking this case fine,
but the error message was not clear.

This commit improves the error message here,
to clearly show:
- the workload version per unit
- interpreted openstack release per unit
- that the information was retrieved from
  the unit workload information from juju

This should help the user to start debugging the issue.

Fixes: #487