Allow users redirect cou output for a file (#636)
* Allow users redirect cou output for a file

sys.stdout.isatty() is a smart module that detect if the command
used is going to run in the shell or redirected.

With the class SmartHalo If it's detected that is a non-tty,
hallo is not used and will just print messages.

Now it's possible to run something like cou plan > my_plan.txt

Closes: #436

* - add spinner_id for the SmartHalo